Refactor is_X_like, better invariant checking for SymInt overload
Add is_symint_like, by way of is_base_ty_like which generalizes
the pattern for is_tensor_like and is_generator_like.  Now that
we can query if a signature contains a SymInt, we can enforce that
you must name the overload with SymInt if the signature contains
SymInt.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: d398b0cd06310876a87b89f56c5ebfca40158fe5
Pull Request resolved: https://github.com/pytorch/pytorch/pull/83668